Max number of concurrent bgtask is defined.
* Running to many BGTask concurrently may cause unexpected problem.
So, max number is defined.
* other misc.
  - temp files are cleaned at the begining of app.
  - interface for RTTask is simplified.

+++ src/free/yhc/feeder/model/Utils.java
@@ -411,14 +411,15 @@ public class Utils {
     // return : false(fail to full-delete)
     public static boolean
     removeFileRecursive(File f, boolean bDeleteMe) {
+        boolean ret = true;
         if (f.isDirectory()) {
             for (File c : f.listFiles())
                 if (!removeFileRecursive(c, true))
-                    return false;
+                    ret = false;
         }
-        if (bDeleteMe)
+        if (ret && bDeleteMe)
             return f.delete();
-        return true;
+        return ret;
     }
 
     public static String
